Mark LADSPA_Descriptor *descriptor as static
The descriptor we return when
    const LADSPA_Descriptor *ladspa_descriptor(unsigned long Index)
is called and that we fill with life in init() actually needs to be
_our_ plugin descriptor. descriptor being a global variable with
external linkage means that we may refer to a variable from some other
loaded library, leading to segfaults or other unexpected behavior.

Marking it as static causes internal linkage and hence makes sure that
it actually refers to the variable we mean.

Fixes #4