Don't expect control ports to point to valid data in connectPortToFilter
connectPortToFilter() receives pointers that _will_ point to valid
control values when run() is called. At the time when
connectPortToFilter() is called, the values behind these pointers may be
uninitialized or invalid. Also, hosts may not call connectPortToFilter()
again when changing values behind these pointers.

To deal with this, just store the pointers and evaluate their values at
the beginning of the run() function. To save us from recalculating our
filter parameters on every call of run(), store the previous control
values and only act on changes.

Also enforce boundaries on the frequency control. Before this change we
accidentally created a filter with frequency set to 0 when this plugin
was used in Ardour, causing horrible audio pops.

Fixes #5